Improve ETL of coded facts. Add codesystem to terminology mapping.
Instead of calling per-row dimension lookup functions, the most expensive
lookups are now replaced with joining the dimensions in the range table
instead.  For the sake of loose coupling, the API functions xxx2set_nk are
still used, but a bit rewritten to be better performing, at the loss of
generality.

This commit adds the codesystem to the terminology mapping table.  Todo is
adding the codesystem version.