Prevent double counting of facts, by linking the provider to only the author.
Previously, each observation event was created twice in the data warehouse,
since the synthetic dataset contains for each observation both the author and
data enterer participations.  A solution is to introduce a bridge table between
dim_provider and the fact tables, so all recorded participations of acts can be
transformed to the data warehouse, at the cost of increased query complexity.
Since the author is arguably the most important provider reference, the current
surrogate key to provider can be kept, so queries that only need the author, do
not need to join the bridge table.  In that case, the current sk to the
provider table must contain only author participations, if any.  This commit is
exactly that change.  The introduction of the bridge table to include more
participation types is left for the future.

The remaining part of this commit consists of copying the recent speed
improvements of the coded observations transformation to the measured values
observation transformation, as well as some minor code cleanups.