Switch to Bashton image since the official CentOS does not recognize storage.
Also switch shax script to use name instead of ip, since the output of
euca-describe-instance differs between ami that do or do not need registration,
and this difference occurs after the name but before the ip number.  Also it is
better to have amazonaws names in known_hosts since that is easier to clean up
than based on ip number alone.