Connect loaders to dwh during setup
Added dwh host as optional setup parameter. If set (as second parameter
to start-group.sh) no dwh instance is started. dwh host is propagated to
the other setup scripts.

Start and connect loader for each pond database (based on number of CPUs).

Start xfm instances based on the number of CPUs.

dwh database now listens on all network interfaces so the loaders
(running on another node) can connect to it.