Add message consume threshold to avoid overflow
Now that the generator is publishing to its own broker we can let it hit
its watermarks, but on "our" other broker the ingress component is still
a lot faster then the loader, so eventually it will hit its watermarks
as the messages are coming in faster then they go out. In this commit a
threshold is introduced that stops consuming (or more precisely caching
the message instead of processing it and therefor no ack is sent), so
the downstream components can catch up.

The threshold is calculated by periodically requesting the cumulative
queue message count of the ingress-hl7v3 and ingress-fhir queues on the broker.

The period of checking the message count (in ms), and the threshold of the combined
number of messages in the transform-hl7v3 and transform-sql can be set in properties.