Use transactions instead of publisher confirms
Now assumes a single thread from upstream to downstream. This is
currently not true for flow control (which is now broken anyway as AUTO ack
mode is used; we could see if starting/stopping the inbound adapter
could be used instead). Also not true for the loader aggregator reaper
which runs in a separate thread.

The transaction attribute (default) now does not rollback on checked
exceptions; need to use a properly configured rule based transaction
attribute.