Add instances to same placement group
Placement groups in EC2 are used for clusters and allow a low-latency,
10Gbps network.

- Modify instance types for placement group compatibility
- Change rabbitmq message store dir to ephemeral for more space

Note that ephemeral storage is meant for temporary storage which is fine
for AXLE, but for other deployments we should move to EBS for example.

- Enforce same availability zone for instances

ec2 api tools do not allow to set the zone or query for the value, so we
found out by trial and error. Zone for instances now fixed to current
placement group's availability zone.

- Increase max connection to lake to allow more loaders
- Add message persistence as configuration option
- Add patientPerson identifiers to patient

After some discussions it was decided to copy identifiers to both the
patientPerson (entity) and patient (role) because it is not possible to
determine which should go where (without matching the OIDs). The special
"polis nummer" identity (matched on having a period element) is left
untouched (i.e. in the patientPerson).