Enable inheritance on the lake for non-columnar relations.
Previously on the lake all RIM inheritance was disabled, to allow columnar
storage and partitioning.  For entity resolution, being able to update a RIM
inheritance root relation is very useful, for instance when a new Person is
added to a cluster, all Role foreign keys to this cluster must be updated.  So
we want RIM inheritance for the Role hierarchy and Participation.  As only
Observation is configured to use columnar storage, enable RIM inheritance for
all other relations again.

As a result, currently only Observation and its two direct childs do not
inherit from Act.