Update generated .bashrc to unset system wide JAVA_HOME.
The Mavan Scala compilation of the cda-generator needs JDK1.7, which it will
get from /usr/bin/java.  Some CentOS versions carry a JAVA_HOME that can be set
to a different, possibly incompatible, version of the JDK.  To workaround this
issue, the user's .bashrc is updated to unset the system JAVA_HOME.  As a
result, the maven will then use the java version as indicated by the java in
the path, which will be set to OpenJDK 1.7 by the setup script, which is the
correct version for the cda-generator scala program.