Add Participation_in table and change Entity Resolution accordingly.
New records are collected in Participation_in. (next commit).

Records in Participation are foreign keys of Role records. Thus a record in
Participation is related to one Role. A Role record is resolved only once, and
then Participation must be updated to reflect changes. To keep these updates
small, Participations are loaded into Participation_in. At resolution of Role,
the foreign key records in the associated Participation_in table are moved to
Participation.

The next commit contains the chance to pond_upload to load Participation in
Participation_in.