Switch to jsonb document store for retinopathy dataset base values source.
This change also introduces the use of TABLESAMPLE on the Patient table, so the
returned dataset can be used to 'get a taste of the data', rather than the
complete dataset.  The sampling percentage depends on the number of patients in
the database, and it is calculated so the returned sample is in the order of 30
patients.  The current script uses a fixed sampling seed, so repetitive calls
return the same data, which is useful for benchmarking and validating results.